+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [[2.6.0](https://github.com/scouting-nl/scouting-openid-connect/releases/tag/2.6.0)] - 19/08/2026
+## [[2.6.0](https://github.com/scouting-nl/scouting-openid-connect/releases/tag/2.6.0)] - 13/08/2026
 
 - Refactored the plugin's PHP and JavaScript codebase for improved readability and consistency.
 - Standardized source file names and the internal plugin structure.